Fix the sporadically failing CLITests (#486)
The CLI tests occasionally fail because the test assertions are too brittle: they expect output like Stream Created And Running to be written to the console immediately after the statement is submitted.

This is mostly, but not always true. It could be that the server is slow, or sometimes we get responses like Queing statement, etc. In these cases, the tests will spuriously fail.

The fix is to wait a little while for the expected output, and also allow for extra statements in the output from the console. This is more flexible and will help keep the test more stable over time.