Add list of actual results to CLITest output when it fails (#494)
This patch adds logging to the `CLITest` so that we know the actual results we got when it fails. It could be that no results are ever returned, or it could be that there is another error and error strings are returned. This patch will learn the cause of the failures.