KSQL-508: add metrics interceptors instead of replacing (#585)
* KSQL-508: add metrics interceptors instead of replacing

Have the physical plan builder add metrics interceptors instead of
replacing the existing ones. It looks at the current setting, and
builds a list of interceptor classes/names out of it. Finally, the
builder appends the metrics collector interceptor to the list.

To test these changes we factored out the last step of KafkaStreams
construction into a factory interface that can be faked out in the test
to intercept the args, which includes the streams config