Fix queries with limit (#615)
* Fix queries with limit

The isReady method of the InputStreamReader used to poll the
input stream for transient queries never gives any indication
that socket was closed by the remote server. Since stream queries
only terminate on error from the server or client side abort,
look out for an error row and stop processing the query when
one is received.