Dont rekey streams for aggregate if group by key (#706)
* Dont rekey streams for aggregate if group by key

When building the kafka streams app for aggregates, dont rekey the
stream if grouping by the key. This avoids expensive repartitioning
in one case where it is not required.

* Move some imports around

* Fix broken test

* Move the decision about whether to group by key into SchemaKStream

Move the decision to group by key or not into SchemaKStream.groupBy. This
function accepts a list of columns for the grouping. If the list is just
the key, then group by key

* Address review comments