Fix some issues with comparison op codegen
Use the right operations for the various operand types:
    - For non-boolean scalars, use (<= && >=) to check for equality.
      == does a reference check rather than a value check. .equals
      requires the argument to be of the same type as the object, so
      equality comparisons between different types dont work. (e.g.
      BIGINT and INT).
    - For boolean scalars, only support == and !=. Implement using
      Boolean.compare
    - For Strings, use .equals for == and !=, and compareTo for other
      operators.
    - For array/map, only support == and != and implement using
      .equals

This patch also includes handling of null operands. The implemented
behaviour is informed by Postgres semantics for null operands. For
every operator other than IS DISTINCT FROM, return false if either
operand is null. Postgres actually returns null, but janino refuses to
accept null for a Boolean, so we return false for now. For
IS DISTINCT FROM, return true if exactly one of the operators is null,
false if they are both null, and defer to != otherwise.