Use equals/compareTo for comparison expressions
Use equals instead of ==/!= operators for comparison expressions. This
ensures that we check value equality for primitive types rather than
reference equality of the wrapper objets. This patch also includes a
change to use compareTo for ordering comparisons. Though not necessary
for primitives, this lets us support ordering comparisons for Strings
which would previously have failed.