Fixes #809 (#810)
General update of the page to bring it inline with the code (and the .md version of the doc).

To make the steps hang together better I've also added an additional step to create an unfiltered join, and use this to build the pageviews_regions later. Which makes more sense than making a 'view count by region and gender' table from a filtered stream that only contains one gender.