WIP: Key constraints documentation
First pass at key constraints doc. Explains that the key should be
STRING/VARCHAR and match the message key. We could expand this to
other data types so long as the user is sure that the key is the
utf-8 encoded representation of the value, but that may be confusing.