Add _command_topic as the default suffix for the command topic.
Currently, the command topic is composed of the following:
"_confluent-ksql-<ksql.service.id>".

By default, this means the command topic is "_confluent-ksql-ksql_".
This isn't a very helpful name for the command topic.

This patch adds a "_command-topic" suffix to the command topic. This
menas that the default command topic would be
"_confluent-ksql-ksql_command_topic".

If you set the `ksql.service.id` to be `production` then your command
topic would be `_confluent-ksql-production_command_topic`.

Seems like a more operator friendly name for a very important topic.