Fix clickstream instructions
Fix some issues with the clickstream instructions. From 4.1 onwards
ksql is included with cp, so we don't need a separate step for
starting up the ksql server. We also just need to run 'ksql' to
start the cli.

This patch also removes references to the local quickstart, which
has not been working for the last couple releases.