Return proper errors from the HTTP server endpoints (#1130)
* Return proper errors from the HTTP server endpoints

This patch updates the rest endpoints to return errors with proper status
codes, and with error messages consistent with other CP projects (e.g.
schema registry, kafka-rest, etc).

On the error message front, instead of defining our own error message type,
inherit from the common ErrorMessage class form rest-utils and add additional
fields that Ksql uses. KsqlErrorMessage defines a basic ksql error message,
which adds a stack trace to ErrorMessage. KsqlStatementErrorMessage is used
to return errors processing individual statements on the /ksql endpoint.
Previously we would return a KsqlEntityList in this case, with the error as
one of the entities. This is hard to reconcile when using non-200 status codes,
since the body could also have a standard error message for such responses.
KsqlErrorMessage includes the statement that hit the error, and a list of
successfully processed entities. Finally, QueryEndpointErrorMessages is a
special type of KsqlStatementErrorMessage for the case where a client issues
a bare query to the /ksql endpoint.

The endpoints have been updated to return proper status codes. The mechanism
for doing so is to either return a Response with the appropriate status code,
or to raise a KsqlRestException that contains the desired response. The latter
was included to make it easier to return responses from functions called by the
first-level handler funciton.

All the possible errors are defined and constructed by the Errors class

* Add exception mapper test

* Fix some review comments

* Fix broken test