Add versioning to the API (#1151)
- Add versions to the HTTP API. Just like other CP services, the KSQL HTTP API lets
the client specify the requested version in the Accept header. The current version is
1, specified as "Accept: application/vnd.ksql.v1+json". application/json defaults to v1.

- Add support for versioning to the websockets query endpoint. For the WebSockets
endpoint the client specifies the version as a URL query parameter. Also adds a unit
test for the websockets endpoint that tests the success path as well as error handling
for invalid versions.

- The CLI checks for invalid version errors. This lets us phase out API versions if we
need to later on.

- Some cleanup of the KsqlRestClient tests, which were split between the cli and
ksql-rest-app package

Fixes (#1151)