SourceDescription cleanup + listing with descriptions (#1147)
* Entity cleanup and listing with descriptions

This patch decouples source and query descriptions. SourceDescription
now describes a single data source (stream or table). The fields specific
to queries were pulled out into QueryDescription. QueryDescription also
includes the sources read from and written to.

This patch also makes it so that SourceDescription and QueryDescription are
not KsqlEntities. A KsqlEntity should be used to return the result of executing
a ksql statement (we should probably also change that name at some point, but
this patch refrains from doing so for the sake of keeping the diff manageable.).
With this change we can return a SourceDescription as part of a composition that
itself is a KsqlEntity.

Finally, add support for a DESCRIPTIONS flag to SHOW, which returns descriptions
instead of the more concise listing objects.

* Fix copypasta

* Review feedback

* Fix broken test

* Fix checkstyle

* Use fields in equals