Add struct support for avro by using Connect Converter API
Now that we're moving towards using Connect's data types as KSQL's internal
representation for records, we can use the Connect Converters to (de)serialize
records. This patch implements a generic Connector-based serializers/deserializer,
and switches KsqlAvroTopicSerde to use that serializer/deserializer, configured
with an Avro converter.

We still require a thin translation layer in KSQL to handle case- insensitivity
, and some of our current limits (e.g. no support for binary fields). This translation
layer is implemented by ConnectSchemaTranslator and ConnectDataTranslator.

Finally, Connect Structs validate that fields are optional when assigning null. This
patch includes a change to make all struct fields optional. In order to switch to using
Struct instead of GenericRow we should just make all our fields optional, but that can
come in a later patch.