Wrap api query id strings in an class (#1370)
Adds an EntityQueryId class to represent Query IDs in the API. We dont
use the existing QueryId class as its fairly decoupled from the API and
should be free to change as we evolve how query ids are generated. On the
other hand the representation of the id in the API must not change.