Do schema inference before writing to output topic (#1140)
* Do schema inference before writing to output topic

The previous change to separate validation form execution moved the schema
inference step into the validation phase. Do this as part of the execution
phase as well. This ensures that we write the inferred schema into the cmd
topic and use that rather than doing schema inference from the engine, which
can change over time

* Address review feedback

* Fix some merge issues