Schema Translation Test + Nested Support in Query Validation Test (#1434)
* WIP: Schema Translation Test + Nested Support in Query Validation Test

Refactor the QueryTranslationTest into EndToEndEngineTest and QueryTranslationTest.
EndToEndEngineTest contains the support classes for describing a test case, piping data
into the Streams app, and validating the result. QueryTranslationTest implements test
discovery for the query validation tests.

Extend EndToEndEngineTest to support arbitrary record value types. Record now stores the value
as Object, and returns a Supplier interface for getting a Serde given a SchemaRegistryClient.

Add support for serializing/deserializing to/from a generic record spec format. Tests that are built
on EndToEndEngineTest can use the generic format to define inputs and expected outputs. This
Is useful for defining nested rows, for example. For now this format is just the json representation
of a row.

Adds a schema translation test (SchemaTranslationTest. This test can be used to validate that
KSQL loads avro data formatted with a given schema correctly. The test is built on the EndToEndEngineTest. Each test case describes an avro schema under test, and an optional
list of expected inputs and outputs. The test creates a source stream that uses the schema and
runs a simple projection of all the contained fields. If no input/output records are specified, random
input records are generated and the expected output should match the input.

* fix some merge problems

* Apply review feedback

* Remove copied parse calls

* Address more review feedback

* More review feedback

* Add complex fields to the basic avro tets

* Fix rebase problems