Extend the query translation test to support parametrized format
This patch allows query translation tests to be parametrized on the value
format, by specifying the format as '{FORMAT}' and providing a 'format'
field in the test spec with a list of formats to test.

The schema for avro format is computed by inferring the schema from a
ddl statement.