Write configs to command topic and read back before building queries (#1479)
* Write configs to command topic and read back before building queries

This patch includes a few prerequisite changes to KsqlConfig to allow writing
configs to the command topic, and reading them back with the correct defaults.
It also includes changes to the command store and statement executor to actually
write the configs out and read them back for use when executing queries.

KsqlConfig has been updated to fully resolve all the streams configs that
are provided to it. Previously we would only check for configs with the right
prefix and/or name. The main problem that resolving the configs solves is
leaking configs that contain sensitive values, such as private key file passwords.
ConfigDef supports obscuring sensistive values from the convertToString method,
but the value needs to be resolved to the correct type first. Resolution is a bit
hacky. In order to access the underlying types we have to instantiate dummy
instances of the config classes (e.g. ConsumerConfig). An alternative solution
was to whitelist configs that we could return from the API or write to kafka.
The implemented solution was chosen because it will work even as new configs are
added.

This patch adds a new type of config def to KsqlConfig called
CompatibiltyBreakingConfigDef. CompatibiltyBreakingConfigDef allows specifying
2 defaults, a current and an old. The current default is used when writing to
the command topic. The old default is used when reading. This way, a config can
be added that takes a new default for new queries and another default for existing
queries. KsqlConfig has a static list of CompatibiltyBreakingConfigDefs. It uses
the list to specify which configs need to be preserved when merging a query's
original config with the current config.

Finally, this patch includes changes to the command store and statement executor
to write out properties in commmands, and read/merge them with the current config
before running queries.

* fix some borked tests

* Fix review feedback

* Fix some rebase issues

* Address review feedback

* Fix some broken tests