Add a schema inference test (#1529)
* Add a schema inference test

This patch adds a test for schema inference. The test ensures that
schemas are inferred correctly and that the correct statement is
produced.

To facillitate the test, the code for inference was moved into a new
class called StatementWithInferredSchema. StatementWithInferredSchema
has two members, the statement and the statement text, and a static
function for inferring the schema from a statement that maybe doesn't
have a schema.

This patch also includes fixes for some issues found along the way
    - Don't fail statements for which the backing topic has a schema
      with types that are not supported by KSQL. Instead, log an error
      and omit those fields from the schema.
    - Bug in the schema translator for handling null names
    - Bug in SqlFormatter for handling statements with empty schemas

* Fix checkstyle

* Review feedback

* Fix review feedback

* merge problems

* Add missing file