Specify a namespace for avro schemas
The avro serializer should provide a namespace in the avro schemas that it generates.
This ensures that users can compile the schema into a class that can be imported from
their own java code. Without the namespace, the generated code is placed in the default
package, and cannot be imported.