Use a thread-local avro serdes for running queries (#1749)
* Use a thread-local avro serdes for running queries

This patch changes running queries to use a thread-local avro serde. The avro
serde has always had some synchronization overhead due to the schema cache in
the schema registry client. The change to avro converter adds to that overhead
with its own caches. Accessing these caches is particularly expensive since it
hashes the whole schema while synchronized.

We use ThreadLocal to make the serdes thread-local, since streams doesnt have
an interface for passing a supplier that can be used to instantiate a serde per
thread. This is implemented using a small utility library under the tls package.

Finally, this patch changes the engine to pass around a schema registry supplier
instead of a schema registry, so each thread can create its own.

* say no to inheritance

* Address review feedback

* Review feedback

* checkstyle

* Fix broken test