CLI test flakiness and clean-up
The CLI integration test is a bit flaky. It issues commands and then
polls the printed output for the desired result. However, some commands
return intermediate statuses if they time out, and of course the CLI
does not itself poll the status and print updates to the console. So
the test's poll loop will never pass if the cli gets an intermediate
status back. This patch changes the test to accept intermediate statuses
for CLI output (which is valid behaviour for the CLI). For setup operations
that should block on a terminal status, the test uses the http api instead.

This patch also cleans up the test a bit:

Firstly, we no longer run poll loops for CLI outputs (except for tests that
exercise the SELECT interface without a limit). Instead, we just wait for the
cli to return and then validate that output.

Don't track ordering in the TestResult object, which is a collection of printed
rows for tables. Instead, just leave that as a collection and let each test case
validate ordering as needed.

Adds a few matchers to make it easy to validate the printed rows.