Fix some synchronization problems with KsqlResource and StatementExec… (#1856)
KsqlResource and StatementExecutor are not synchronized properly:
    - KsqlResource registers for a future after writing commands to the
      command topic. By this time, the StatementExecutor may have already
      picked up the command and missed the future created by KsqlResource
    - If multiple threads try to perform the same operation around the same
      time, they could both end up listening on the same future and one of
      them will then return an incorrect response. This can also happen
      across servers, but this PR does not attempt to solve that (will file
      a follow-up to cover that case, since its a bit more controvertial
      as it requires a change to the command messages).
    - The id -> status map is not synchronized, but is accessed by multiple
      threads

To fix these problems, we've done the following:

KsqlResource/CommandStore now register for futures before writing to the
command topic. All communication between the StatementExecutor and
KsqlResource happens over this future.

If another request comes in and tries to register for a future for the
same logical operation, we fail the request. As stated above, this could
also happen across servers, in which case the implemented check won't help.
To fix that, we'll need to add a "cookie" to the command topic that
StatementExecutor can look up to find the future that KsqlResource is waiting
on. This way we can make a strong correlation between a command being executed
and a waiting server thread. That change will come in a follow-on PR (at which
point we could get rid of the check).

Use a ConcurrentHashMap for the id -> status map.