Add docs that describe metrics returned by describe extended (#1956)
This patch adds docs that provide info on the metrics returned by describe. The
naming on the failed messages isn't a typo. The only failure we count today is
deserialization failure, and it updates a metric without the the consumer- prefix.
We should fix that, but not on the 5.0.0-post branch