Retry commands on unexpected exceptions
This patch changes CommandRunner to retry commands when it hits an error that
may not be hit by other KSQL servers. If it just ignores such errors, the servers
in a cluster could become inconsistent. The number of retries is controlled by a
config that defaults to INT_MAX. We set it to a lower number for tests to ensure
we dont accidentally hang while running tests.