Enable support for avro unions (#2061)
This patch enables support for avro unions. Unions have always been supported
by connect converters. The converter converts a union to a struct, with a field
for each possible type of the value in the union. All the fields are null, except
for the field corresponding to the type of the value, which is set to the value.