Fix some bugs in recovery logic (#2143)
This patch fixes some issues with our recovery logic by simplifying the
recovery algorithm. The new algorithm is as follows:
    - Initialize a list recoveredQueries
    - For each command in the command topic:
        - execute the command
        - if the command creates a query, add it to recoveredQueries, but
          dont start it.
    - For every query in recoveredQueries:
        - if the query is still in the engine's set of queries, start it.

For 4.1 DROP commands, terminate the query writing into the source being
dropped, if any. We recognize 4.1 DROP commands by looking for the
originalProperties field.

Adds a new test, called recovery test, that runs sequences of statements
in execution and replay mode, and verifies that the resulting engine has
the same metastore and queries.