Use the KafkaMetrics metricValue call to get metric values from KsqlEngineMetrics
This patch fixes a race that happens when KsqlEngineMetrics tries to sample query metrics.
KsqlEngineMetrics wakes up every second and samples metrics by calling TopicSensors.SensorMetrics.value.
This function gets tthe measurable from the metric to get the value. This is not a thread-safe way of
getting the metrics value, and can result in a ConcurrentModificationException. We should instead call
KafkaMetrics.metricValue()

The result of the above race is that jmx metrics are never updated once the exception is thrown. In
general, it doesn't make sense to stop sampling/updating metrics forever on a possibly transient exception.
So this patch also has a change to catch-and-log exceptions from the metric updater.

This patch also enables the mock-maker-inline feature in Mockito, which tells mockito to set up mocks
such that final classes and methods can be mocked. This is an incubating feature. Hopefully it continues
to work, but we may need to revisit down the line.
(see https://github.com/mockito/mockito/wiki/What's-new-in-Mockito-2#mock-the-unmockable-opt-in-mocking-of-final-classesmethods)