Add processing logger and first set of log msgs for deserialization errors
This patch adds a package for emitting diagnostic record proessing logs from
ksql as queries are processing messages. Loggers are created using
ProcessingLoggerFactory. The logger can log structured messages at various
log levels. Messages are represented as connect SchemaAndValues. This patch
also includes a factory class for creating various types of messages that
ksql will emit.

A lot of the code here is for the plumbing of passing around the info
needed to create loggers with the right names. The loggers are named in
a hierarchy of the form:

<query id>.<logical plan node>.<operation>.<op-secific hierarchy>

   query id: the id of the query producing the log message
   logical plan node: the name of the node in the logical plan that the
                      query was executing when the log message was
                      emitted.
   operation: describes the operation being performed (e.g. repartition
              or source)