Enable streams optimizations for KSQL queries (#2077)
* Enable streams optimizations for KSQL queries

This patch updates the engine to enable optimizations for KSQL queries
by default.

The bulk of the changes are to ensure that queries are built compatibly
with optimizations enabled, and that existing queries with optimizations
disabled don't break. There are 2 parts to ensuring compatibility:

Firstly, we need make sure we choose the same optimization setting
every time we build a query. To accomplish this, this patch adds
compatibility-breaking _streams_ configs, and includes the
optimization config in that set.

Secondly, we need to choose our own names for internal topics.

APIs for setting names for internal topics were added in AK 2.1. We've
wrapped the APIs in an interface to make them more easily mockable. This
interface, called StreamsStatics canbe found in a new package, called
io.confluent.ksql.streams.

However, for older queries built without optimizations we should use the
older naming conventions. To handle this logic, we've added factories for
the Streams objects that take topic names (Grouped, Joined, Materialized),
and put the logic for switching the names based on the optimization config
in those factories.

* Add back grouped stream test

* Review comments and checkstyle

* use the provided optimization config

* Add test

* Address some review comments

* review feedback

* Add separate config for topic naming

* missed test