switch `KsqlEngine.execute` to accept single statement.
Fixes: #2202

The issue with multiple statements is that if a later statement fails then the effects of the former ones on the metastore are not rolled back and queries are left zombied.

This change switches to accepting only a single statement. This removes the issue.

The normal flow should be:

```
// Parse:
statements = engine.parse(sqlText);

// Try to execute all of them using a separate metastore:
engine.tryExecute(statements);

// Now we know they are all valid we feed them into the engine one by one:
statements.forEach(stmt -> engine.execute(stmt));
```