Hopefully fix flaky WindowingIntTest
This patch fixes an issue with WindowingIntTest where we check the
size of the returned record set on each call to verifyAvailableUniqueRow.
In the case where the first call doesn't return all the rows, this causes
the test to fail prematurely. This patch fixes this issue by skipping the
size check