This patch adds some more query processing log messages, and lays some foundations for future work in this area:
The main new functionality is to log errors to the record processing log when a
projection or a filter errors out. The error is logged using a new record type,
called a "record processing error", that contains a description of the error and
the row that triggered the error serialized as a json string.

This patch also consolidates streams operation naming (for naming internal topics)
and record processing logger naming into a single pattern for generating unique
hierarchical names in an execution plan. This is implemented as QueryContext and
QueryContext.Builder. The Builder is initialized with the Query ID, and context
strings can be pushed onto it as we build the query plan. The final context is
returned by getQueryContext, and is used to construct streams op names and record
processing logger prefixes. In the future, we can use this to identify steps in the
query execution plan when returning it from `DESCRIBE EXTENDED`.

Add the record processing logger name to the query execution plan. This way, users
can corellate steps in the execution plan to log messages, and once trace logging is
enabled, use the logger name to enable more detailed logging for specific steps of
their queries. For now, this execution plan is returned as a summary string. In a
follow up, I plan to include a more detailed json representation with details about
the schema at each step, and what each step is doing (e.g. the select expressions for
project steps).

Minor changes:
- Split up the ProcessingLogMessageFactory into a class that defines the schema, and
  factories for the various modules.
- Move json mapper / struct serialization module to common so the engine can use it
  to serialize rows for the record processing log.
- Move the various streams factories for grouped, joined, materialized, etc into a
  wrapping class (StreamsFactories) to keep the parameter count for SchemaKStream
  within limits.
- Switched SchemaKStream to mockito
- Simplified the SqlPredicate test (similar to the SelectValueMapperTest)
- Add the expression responsible for the error in project/filter error messages