There can be only one... (#2351)
This PR is all about refactoring the KsqlEngine to allow headless and interactive modes to use the same code paths, (a.k.a. unification). This requires that the tryExecute and execute methods of the engine follow virtually the same code path.

Gone is the updateMetaStore flag that was passed around. In its place, the KsqlEngine now has the concept of an ExecutionContext. This wraps a MetaSore, ServiceContext and QueryEngine, amongst others. There is the primaryContext, used when actually executing statements for real. From this the engine can create a try execution context. The try context:

copies the primary context's meta store
has its own QueryEngine, so there's no issue with query ids being messed up
has a TryServiceContext which massively reduces what can be done with external services.
The engine makes use of this try execution context within both the parse and tryExecute methods, ensuring that these methods can neither mutate the primary metastore or remove services.
The TryServiceContext internally makes use of try variants of the client interfaces for Kafka and SR. Some of these allow no operations, (except close), while others allow a very minimal set of read-only operations. The exception to this is TryKafkaTopicClient which will also track any createTopic calls, (though not actually create the topics), and use this tracked information to later allow calls to isTopicExists and describeTopics to succeed. In all of this try clients, only the calls that are currently needed are implemented. Everything else throws UnsupportedOperationException. As an when other functionality is needed, these can be implemented.

For this all to work, the way tryExecute worked needed to change. Previously, it would preprocess all statements, then build all logical plans, then finally physical plans. To avoid issues with placeholders in the metastore, tryExecute now processes each statement fully before moving onto the next, e.g. it will preprocess, build logical, then physical plans for a statement before moving onto the next.

KsqlEngine has been further refactored to pull parse specific functionality into an inner EngineParser class, and pull [try]execute specific functionality into EngineExecutor. This, IMHO, makes the code more readable.