Processing log stream + docs (#2387)
Processing log stream + docs

This PR adds supports for setting up a processing log stream automatically,
and includes some documentation for the whole feature.

Creation of the log stream is configurable using the prop
'ksql.processing.log.stream.auto.create'.
The log stream can only be created automatically in interactive mode. This is
because the stream's schema is inferred and may evolve over time. In interactive
mode the schema of the stream is stored in the command topic. In headless mode,
the schema could evolve in ways that break downstream queries. Users can still
use the log stream in headless mode, they just have to write the ddl explicitly
into their query files.

We also only support automatic log stream creation for new clusters. This is
because its tricky to detect if the stream was previously created and then
dropped.

This patch also adds support for automatic creation of the processing log topic.

Finally, this patch adds some documentation on usage of the processing log.