Add config to disable logging row data from processing log + config docs (#2419)
* Add config to disable logging row data from processing log

This patch adds a config to disable logging actual row data from
the processing log. The config will be defaulted to true. The
reasoning for this is that its dangerous from a security pov to
log row data, so users should be forced to do configure this
behaviour explicitly.

From an implementation POV, this is used by making the processing
log config a global, and accessing it from the various message
factories. We should also consolidate all the processing-log related
code to its own module, but thats deferred to a follow-on to keep
this PR small.

* Add the processing log docs to the server config reference

* Address review feedback

* Review feedback

* Apply suggestions from code review

Co-Authored-By: rodesai <desai.p.rohan@gmail.com>