Persist configs standalone 5.2.x (#2464)
This patch changes the standalone mode executor to use configs persisted in a kafka
topic, rather than just the configs present in the config file. The main advantage
here is that new standalone apps can use the latest defaults without any special
configuration, and continue to work across upgrades even if the default configs
change in an incompatible way. The functionality is implemented behind an interface
called ConfigStore, that has a single method called getKsqlConfig. The implementation,
KafkaConfigStore works by reading configs from a designated kafka topic and merging
them with the configs provided at runtime. If no configs are present in the topic
KafkaConfigStore writes the current configs.