Adds a name to map schemas (#2504)
* Adds a name to map schemas

This patch adds names to map schemas when serializing avro. Internally,
the avro converter will use this name to name the map entry type. This
fixes the issue where, when a schema has multiple maps with different
types, avro errors out because the same name is used for different
map entry record schemas

* fix qvt

* guard the schema name change with a compatibility-break config

* fix build

* fix tests