Change archetype integration test goal to 'compile'
The surefire plugin isn't playing nicely with the archetype plugin.

The archetype plugin drops generated test classes in a nested directory, i.e.

    test-classes/projects/it0/project/.../com/acme/ksql/functions/ReverseUdfTests

However, the surefire plugin expects the test to be in the following directory, based on the generated package name of
the archetype instance:

    test-classes/com/acme/ksql/functions/ReverseUdfTests)

Therefore, the verify step fails. Changing the goal from 'verify' to 'compile' still gives us some form of verification,
but unfortunately doesn't allow us to run the example UDF tests.