Add an additional test to key fields to cover the case where the source's key field, say 'K', is aliased in the projection, e.g. `K as A`, and another field is aliased in the project to match the source's key field, e.g. `X as K`.
Currently, in these situation the key field of the new source would incorrectly be `K`, not `A`.  This introduces a new complication in that downstream queries that `REPARTITION BY` or `GROUP BY` `K` should do a repartition, but won't and so won't work correctly, and any that `REPARTITION BY` or `GROUP BY` `A` will do an unnecessary repartition set.

I've also documented implications of current key field management in `key-field.json`, as much for my benefit as for others.