Add concept of `originalConfig` to `QueryTranslationTest`.
Previously, QTT only supported `properties` being set in the test cases. This is passed through in the same way as a user issuing a `SET x=y` style command.

Unfortunately, `properties` is not sufficient to test what happens in the code should a property be set either in the main config file, or the server is recovering and building up the config to run the command with from the main config file and the `originalConfig` stored in the command in the command topic - this latter set of properties is used to set the `CompatibilityBreakingConfig`.

This PR adds the concept of `originalConfig` alongside the existing `properties` to allow test cases to better test backwards compatibility.

Without this PR it is possible to write a test case with compatibility breaking config set in `properties` that pass, but which would not actually work in reality as the code will get such compatibility breaking config through the `ksqlConfig`, not `propertyOverrides`.