Add cause for errors logged to processing log (#2662) (#2711)
* Add cause for errors logged to processing log

Java exceptions can have a chain of causing exceptions that contain valuable
context about why an exception was thrown. KSQL has an ErrorUtils utility
class for building error messages from the cause chain. In this patch, we
add a utility function to that class which returns the cause error msg chain
as a list, and include this list in the record processing log as a new field
for deserialization and processing errors.

* Docs