Don't do unnecessary rebuild of sink topic serde for avro (#2698)
* Don't do unnecessary rebuild of sink topic serde for avro

For some reason KsqlStructuredDataOutputNode is rebuilding the sink topic serde
for queries that write avro. There is no reason to do this, and in the case of
INSERT INTO from a json source actuall results in an NPE, because the schema name
isn't set in the properties. Instead of doing this rebuild, we should just use
the serde the analyzer computed for us.

* review feedback