Switch metric names to a format that can be consumed by datadog.
The previous names produced a statically named beans, and embedded the
service ID in the bean attribute. With this patch, we embed the service
id in the bean name, and use statically named attributes.

In this patch, we publish under both the new scheme and the old scheme.
The old scheme will be removed in a later ksql version.