Add support for checking external preconditions while still serving
Refactor the server setup to first create all the resources and start
accepting requests before initializing and replaying the command topic.
While the command topic is being initialized/replayed, KSQL requests
will return an error indicating that the server is initializing.

As part of this change, I've consolidated this initializing state and
the terminating state (entered when a TERMINATE request is received)
into a single state machine tracked under the package
io.confluent.ksql.rest.server.state. This package also has a request
filter that ensures that requests are only served when the server is
in READY state.

Finally, this patch adds support for checking preconditions before
initializing the server. The preconditions are provided in config
via the ksql.rest.server.preconditions property.