Improve test coverage of avro serde - ensure correct schemas are registered by serializer - ensure deserializer can handle connect's optional key maps.
Fix bug where schemas of unwrapped types where being registered as unions, i.e. optional. Although this is strictly true, it would be inconsistent with wrapped schemas, which are not an optional record. Basically, the root schema doesn't need to be optional, as `null` is always handled _before_ getting to the avro deserialization.