Switch how we handle implicit columns in sources. The source in the meta-store now has a schema that matches its data.
The implicit and key columns are still added to the value schema, but rather than adding them to the schema of the source stored in the metastore, they are added during the processing of queries, to match the schema of the different parts of the query topology, i.e. the schemas defined in the code now match the schemas of the data at the different points, i.e. source / sink topics do NOT have ROWKEY and ROWTIME in the value schema, these are added by DataSourceNode and later removed by KsqlStructuredDataOutputNode, (well removed if its not a join).