refactor: make sql types available in common module
In the future we want to be able to change `LogicalSchema` to have its own `Field` type. This `Field` type will not use the Connect schema, but instead should use the KSQL type system.  `LogicalSchema` exists in the common module, therefore we need our SQL types available in the common module.

This change moves our SQL types into the common module, and converts the parser's `Type` from a base class to a holder of `SqlType`.

Some renames have occurred as part of this process:

* `SqlType` (enum) -> `SqlBaseType`. As there is now a new `SqlType` base class for all our SQL types.
* `Array` -> `SqlArray`. As this avoids clashes with other Array classes
* `Map` -> `SqlMap`. As this avoids clashes with other Map classes
* `Struct` -> `SqlStruct`. As this avoids clashes with other Struct classes
* `Decimal` -> `SqlDecimal`. As this is inline with the other name changes.
* `PrimitiveType` -> `SqlPrimitiveType`. As this is inline with the other name changes.