refactor: remove topic from metastore api
Now that all the `REGISTERED TOPIC` stuff has been removed from our syntax the relationship between `DataSource` and `KsqlTopic` is one-to-one. This means there is no need to have a `putTopic` and `getTopic` etc on the Metastore API. All operations can be done via the `DataSource`.

This PR strips all the topic handling functionality from the metastore interfaces. `KsqlTopic` is now just used a a field within the data source.

This simplifies the code no end, and there will be a follow up PR to simplify more by removing the duality of `ksqlTopicName` vs `kafkaTopicName` which causes such confusion and, hence, bugs.

To make this change I had to reword `ClusterTerminator` a little to make use of the topics within the data source, otherwise removal was fairly straight forward.