feat[rest server]: add warnings when listing fails with kafka error (#3069)
This patch adds a mechanism for returning warnings in KsqlEntity for
successful api responses. This patch also adds some warnings for source
listings. Now, when a client tries to list or describe a source, and ksql
is unable to get the details of the backing topic from kafka, a result is
returned that includes 0-values for the underlying topic info (partitions,
replicas), and includes response warnings about the issue with Kafka.