fix: check for other sources using a topic before deleting (#3070)
When handling DELETE TOPIC, check that other sources don't use the topic
before deleting. This is an optimistic check - the user could still delete
topics out from under us, and its not safe racing against concurrent statements.
But, we've seen single users kick themselves this way in the past, so this should
help in those cases.