refactor: drop KsqlTopic's KSQL topic name
Historically `KsqlTopic` had too _topic names_:

1. `kafkaTopicName` the name of the topic in Kafka.
1. `ksqlTopicName` the name of the registered topic in KSQL.

This was because previously you could use `REGISTER TOPIC` syntax to register a Kafka topic in KSQL with a different name, and then use that registered topic in C* statements.

This duality of topic names caused several bugs and much confusion. As we've now dropped the `REGISTER TOPIC` syntax the `ksqlTopicName` always matches the name of the stream/table the `KsqlTopic` is associated with, and that association is now always 1-2-1.

Hence, we can now safely drop `ksqlTopicName` to remove the confusion.