refactor: add VisitParentExpressionVisitor
This patch adds an abstract base visitor for ExpressionVisitor, called
VisitParentExpressionVisitor, which provides a default implementation of
each visit call that just delegates to a visit to the parent type. E.g.
visitBooleanLiteral calls visitLiteral calls visitExpression. This is
basically what the existing AstVisitor does for the AST. This patch also
migrates the expression-only visitors that inherit from AstVisitor to
VisitParentExpressionVisitor